+++ src/optparse.c
@@ -1,5 +1,24 @@
 /* vim: set et ts=4 sw=4 sts=4 fdm=marker syntax=c.doxygen : */
 
+/*
+t64fix - a small tool to correct T64 tape image files
+Copyright (C) 2016  Bas Wassink <b.wassink@ziggo.nl>
+
+This program is free software; you can redistribute it and/or
+modify it under the terms of the GNU General Public License
+as published by the Free Software Foundation; either version 2
+of the License, or (at your option) any later version.
+
+This program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License
+along with this program; if not, write to the Free Software
+Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+*/
+
 /** @file   optparse.c
  *
  * @brief   Simplified command line parser
@@ -24,8 +43,12 @@
 #include <assert.h>
 #include <errno.h>
 
+/* needed for `bool` typedef */
+#include "base.h"
 
 #include "optparse.h"
+
+
 
 
 /** @brief  Initial size of argument list
@@ -149,7 +172,7 @@ static int handle_option(option_decl_t *option, const char *arg)
 
     switch (option->type) {
         case OPT_BOOL:
-            *((int *)(option->value)) = 1;
+            *((bool *)(option->value)) = 1;
             break;
         case OPT_INT:
             if (arg == NULL) {
@@ -160,7 +183,7 @@ static int handle_option(option_decl_t *option, const char *arg)
             /* check for invalid crap */
             if (endptr == arg || errno == ERANGE) {
                 fprintf(stderr,
-                        "%s: failed to convert option argument to int: %s",
+                        "%s: failed to convert option argument to int: '%s'\n",
                         prg_name, arg);
                 return -1;
             }
@@ -200,7 +223,7 @@ static int print_option(option_decl_t *option)
 
 /** @brief  Display usage message and options list
  */
-static void optparse_help(void)
+void optparse_help(void)
 {
     option_decl_t *opt = options;