+++ src/prg.c
@@ -21,6 +21,10 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 
 /** @file   prg.c *
  * @brief   Extracting PRG files
+ *
+ * Functions to extract program (*.prg) files from t64 images. 'Frozen' files
+ * ('frz') are skipped, these are snapshots of the C64S emulator and nobody
+ * in their right mind would use that emulator today.
  */
 
 #include <stdio.h>
@@ -33,6 +37,17 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 
 #include "prg.h"
 
+
+/** @brief  Extract prg file at \a index from \a image
+ *
+ * @param   image   t64 image
+ * @param   index   index in \a image of file to extract
+ * @param   quiet   don't output anything to stdout/stderr
+ *
+ * @todo:   Needs proper PETSCII to ASCII translation
+ *
+ * @return  bool
+ */
 bool prg_extract(const t64_image_t *image, int index, int quiet)
 {
     t64_record_t *record;
@@ -46,6 +61,14 @@ bool prg_extract(const t64_image_t *image, int index, int quiet)
     }
 
     record = image->records + index;
+
+    /* check for memory snapshot */
+    if (record->c64s_ftype > 1 || record->c1541_ftype == 0x00) {
+        if (!quiet) {
+            fprintf(stderr, "t64fix: skipping memory snapshot\n");
+        }
+        return true;
+    }
     /* translate filename */
     for (i = 0; i < T64_REC_FILENAME_LEN; i++) {
         int ch = record->filename[i];
@@ -66,6 +89,34 @@ bool prg_extract(const t64_image_t *image, int index, int quiet)
 }
 
 
+/** @brief  Extract all files
+ *
+ * @param   image   t64 image
+ * @param   quiet   don't output information on stdout/stderr
+ *
+ * @return  bool
+ */
+bool prg_extract_all(const t64_image_t *image, int quiet)
+{
+    int i;
+    int skipped = 0;
 
-
+    for (i = 0; i < image->rec_used; i++) {
+        t64_record_t *record = image->records + i;
+        if (record->c64s_ftype > 1 || record->c1541_ftype == 0x00) {
+            if (!quiet) {
+                printf("t64fix: skipping file %d: memory snapshot\n", i);
+            }
+            skipped++;
+        } else {
+            if (!prg_extract(image, i, quiet)) {
+                return false;
+            }
+        }
+    }
+    if (!quiet) {
+        printf("t64fix: extracted %d files\n", image->rec_used - skipped);
+    }
+    return true;
+}