+++ CHANGES.md
@@ -2,6 +2,17 @@
 
 A list of changes applied to the codebase, started too late as usual ;)
 
+## 2016-07-11
+
+* Add extracting program files with -e/--extract \<index\>, or -x/--extract-all
+  to extract all program files
+
+
+### 2016-07-10
+
+* Add simple command line parser (src/optparse.{c,h})
+* Add quiet mode (-q/--quiet) for use in scripts
+
 
 ### 2016-07-08